// frontend/web/src/services/API/proserve-wb-publishing-api/models/VersionDistribution.ts
export interface VersionDistribution {
    /**
     * Project ID
     * @type {string}
     * @memberof VersionDistribution
     */
    projectId: string;
    /**
     * Product ID
     * @type {string}
     * @memberof VersionDistribution
     */
    productId: string;
    /**
     * Version ID
     * @type {string}
     * @memberof VersionDistribution
     */
    versionId: string;
    /**
     * The version type (RELEASE_CANDIDATE|RELEASED|RESTORED)
     * @type {string}
     * @memberof VersionDistribution
     */
    versionType: string;
    /**
     * AWS account ID where this version is distributed
     * @type {string}
     * @memberof VersionDistribution
     */
    awsAccountId: string;
    /**
     * AMI id of the version distribution in the region
     * @type {string}
     * @memberof VersionDistribution
     */
    copiedAmiId?: string;
    /**
     * Original AMI id of the version distribution
     * @type {string}
     * @memberof VersionDistribution
     */
    originalAmiId?: string;
    /**
     * ECR image digest
     * @type {string}
     * @memberof VersionDistribution
     */
    imageDigest?: string;
    /**
     * ECR image tag
     * @type {string}
     * @memberof VersionDistribution
     */
    imageTag?: string;
    /**
     * Region where this version is distributed
     * @type {string}
     * @memberof VersionDistribution
     */
    region: string;
    /**
     * Stage of the version distribution
     * @type {string}
     * @memberof VersionDistribution
     */
    stage: string;
    /**
     * Status of the version distribution
     * @type {string}
     * @memberof VersionDistribution
     */
    status: string;
    /**
     * Last update date of the product.
     * @type {string}
     * @memberof VersionDistribution
     */
    lastUpdateDate: string;
    /**
     * Name of the version that this version is restored from, if it is restored
     * @type {string}
     * @memberof VersionDistribution
     */
    restoredFromVersionName?: string;
}